Jetpack Compose 탐색
 - RallyNavHost에 NavHost추출
